requirements.txt
playwright==1.44.0
pytest==8.2.0
pytest-playwright==0.5.0
pytest-asyncio==0.23.7
loguru==0.7.2
pyyaml==6.0.1
python-dotenv==1.0.1
